Refactor inicial
